1. remove extra # in cmake files 2. fix a bug in FindANTLR.cmake 3. download antlr jar to build directory instead of the source directory 4. add StorageType to ClpTableHandle 5. use consistent types for filteredRows across different methods 6. remove archiveSource_ from ClpConfig and config docs, use storageType_ in ClpDataSource 7. minor improvements in logs and method signatures 8. rename splitPath to path in ClpConnectorSplit

+++ velox/connectors/clp/search_lib/ClpQueryRunner.cpp
@@ -61,7 +61,7 @@ void ClpQueryRunner::init(
 
 uint64_t ClpQueryRunner::fetchNext(
     size_t numRows,
-    const std::shared_ptr<std::vector<size_t>>& filteredRowIndices) {
+    const std::shared_ptr<std::vector<uint64_t>>& filteredRowIndices) {
   size_t rowsfiltered = 0;
   size_t rowsScanned = 0;
   while (curMessage_ < numMessages_) {
